plotting update
Plot RFs and Plot Timecourses moved over to analysis_chunk, with helper functions in vision_utils. Can now call plot_rfs and plot_timecourses as methods on EITHER the analysis_chunk object or the pipeline object. The analysis chunk object takes a list of noise ids and/or cell types while the pipeline version takes a list of protocol ids and/or cell types and does the conversion to noise_ids before passing that variable into the analysis chunk's plotting functions