Fixed hidden error in parse_data
Parse data called 'out_path' directly instead of self.json_path within the read_write() method for Symphony2Reader. This is a hidden error if you run parse_data.py in the terminal using the argument parser (as initially intended) because out_path is defined globally in that case. But if you instantiate a reader and try to run read_write() with all appropriate inputs separately, outside the terminal, the code will fail. That typo has been fixed.